requirements.txt
Flask
face-recognition
scikit-learn
numpy
pillow